Move some functions out of bitcoinrpc.cpp to rpcnet.cpp
